+++ src/main/java/com/fine/model/stock/ChemicalStockDetail.java
@@ -43,9 +43,9 @@ public class ChemicalStockDetail {
     @TableField("pack_uom")
     private String packUom;
 
-    /** 当前包装数量（整数） */
+    /** 当前包装数量（支持小数） */
     @TableField("pack_count")
-    private Integer packCount;
+    private Double packCount;
 
     /** 标准单位：kg */
     @TableField("std_uom")
